Move ColorMatcher class to it's own file in new namespace `stardew_access.Utils` (advances khanshoaib3/stardew-access#112).
